+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # MyChatAlert
 
+## 2.4.4
+
+- New feature: chat command to ignore names
+  - `/mca ignore {name}`
+- Color customization for printed alerts
+
 ## 2.4.3
 
 - New feature: Time-based duplication filter
@@ -15,7 +21,7 @@
 - New feature: Disable in instance
   - With this option selected, alerts will be toggled off while you are inside an instanced zone (determined via [IsInInstance](https://wow.gamepedia.com/API_IsInInstance))
 - `Escape` key will now close the alert frame
-- New Feature: Customize the printed alert messages
+- New feature: Customize the printed alert messages
   - Color customization coming soon
 
 ## 2.4.2